requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
openai-whisper>=20231117
requests>=2.31.0
Pillow>=10.0.0
pydub>=0.25.1
faster-whisper>=0.6.0
python-dotenv>=1.0.0
